+++ src/main/java/com/spm/ibooking/repository/SeatRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.spm.ibooking.entity.*;
 
 @Repository
-public interface SeatRepository extends JpaRepository<Seat, Long> {
+public interface SeatRepository extends JpaRepository<Seat, Integer> {
     List<Seat> findByRoomId(Integer roomId);
 }
